backend/generate-embedding/requirements.txt
Flask
deepface
pillow
numpy
tf-keras